Fixed merge issues

+++ contracts/minter/src/msg.rs
@@ -12,7 +12,7 @@ pub struct InstantiateMsg {
     pub sg721_code_id: u64,
     pub sg721_instantiate_msg: Sg721InstantiateMsg,
     pub start_time: Option<Expiration>,
-    pub per_address_limit: Option<u64>,
+    pub per_address_limit: Option<u32>,
     pub unit_price: Coin,
     pub whitelist: Option<String>,
 }
